% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_frequencies.R
\name{allele_frequencies}
\alias{allele_frequencies}
\title{Compute allele frequencies per markers and populations}
\usage{
allele_frequencies(
  data,
  verbose = TRUE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
\emph{How to get a tidy data frame ?}
Look into \pkg{genometranslator} \code{\link[genometranslator]{tidy_genome}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
A list with allele frequencies in a data frame in long and wide format,
and a matrix. Local (pop) and global minor allele frequency (MAF) is also computed.
}
\description{
Compute allele frequencies per markers and populations.
Used internally in \href{https://github.com/thierrygosselin/radr}{radr}
and might be of interest for users.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
