% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hwe.R
\name{blacklist_hw}
\alias{blacklist_hw}
\title{blacklist_hw}
\usage{
blacklist_hw(
  x,
  unfiltered.data,
  data.temp,
  hw.pop.threshold,
  path.folder = NULL,
  pop.id.levels,
  verbose = TRUE
)
}
\arguments{
\item{x}{HWE summary grouped by significance threshold.}

\item{unfiltered.data}{Original tidy genomic data.}

\item{data.temp}{Optional data retained outside the HWE calculations.}

\item{hw.pop.threshold}{Number of strata allowed to depart from HWE.}

\item{path.folder}{Output directory. Default: \code{path.folder = NULL}.}

\item{pop.id.levels}{Ordered strata levels.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}
}
\description{
blacklist hw
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
