% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr_dots.R
\name{dots_deprecated_core}
\alias{dots_deprecated_core}
\title{Global deprecated dot-arguments for radr}
\usage{
dots_deprecated_core()
}
\description{
Internal helper returning deprecated argument names for \code{...}.
}
\keyword{internal}
