% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr_dots.R
\name{extract_dots}
\alias{extract_dots}
\title{extract_dots}
\usage{
extract_dots(n, v, env.arg, assign = TRUE, verbose = TRUE)
}
\description{
Extract and assign \code{...} arguments (keepers)
}
\keyword{internal}
