% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_common_markers.R
\name{filter_common_markers}
\alias{filter_common_markers}
\title{Filter common markers between strata}
\usage{
filter_common_markers(
  data,
  filter.common.markers = TRUE,
  fig = FALSE,
  parallel.core = parallel::detectCores() - 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radr:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Use \code{\link[genometranslator]{read_genome}} to import supported formats
and \code{\link[genometranslator]{tidy_genome}} when a tidy table is needed.}

\item{filter.common.markers}{(optional, logical)
Default: \code{filter.common.markers = TRUE}.}

\item{fig}{(optional, logical) \code{fig = TRUE} will produce a ComplexUpset
figure to visualize the number of markers between populations.
Default: \code{fig = FALSE}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = FALSE}.}

\item{...}{(optional) To pass further arguments for fine-tuning the function
and legacy arguments.}
}
\value{
A list with the filtered input, whitelist and blacklist of markers..
}
\description{
The function will filter the markers by keeping only those
in common between all strata
(population or any groupings defined in \code{STRATA} column).

\strong{Filter targets}: SNPs

\strong{Statistics}: strata genotyping rate per SNPs

Used internally in \href{https://github.com/thierrygosselin/radr}{radr}
and might be of interest for users who wants to keep only markers in common.
}
\examples{
\dontrun{
require(SeqArray) # when using gds
common <- radr::filter_common_markers(data = "my.radr.gds.rad", verbose = TRUE)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
