% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fis.R
\name{filter_fis}
\alias{filter_fis}
\title{Fis filter}
\usage{
filter_fis(
  data,
  approach = "haplotype",
  fis.min.threshold,
  fis.max.threshold,
  fis.diff.threshold,
  pop.threshold,
  percent,
  filename,
  verbose = TRUE
)
}
\arguments{
\item{data}{A GDS filename or open \code{SeqVarGDSClass} object.}

\item{approach}{Character. By \code{"SNP"} or by \code{"haplotype"}.
The function will consider the SNP or haplotype statistics to filter the marker.
Default: \code{approach = "haplotype"}.}

\item{fis.min.threshold}{Number.}

\item{fis.max.threshold}{Number.}

\item{fis.diff.threshold}{Number (0 - 1)}

\item{pop.threshold}{Fixed number of pop required to keep the locus.}

\item{percent}{Is the threshold a percentage ? TRUE or FALSE.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the folder created in the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}
}
\description{
Filter markers or loci using within-stratum inbreeding-coefficient
statistics.

\strong{Filter target:} Markers or loci, depending on \code{approach}.
}
\details{
The Fis calculated uses the ratio of averages (1-mean(Ho)/mean(Hs))
and NOT THE AVERAGE OF RATIOS (Nei 1987).
}
\references{
Nei M. (1987) Molecular Evolutionary Genetics. Columbia University Press.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
