% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr.R
\name{generate_filename}
\alias{generate_filename}
\title{Filename radr}
\usage{
generate_filename(
  name.shortcut = NULL,
  path.folder = NULL,
  date = TRUE,
  extension = c("tsv", "gds.rad", "rad", "gds", "gen", "dat", "genind", "genlight",
    "gtypes", "vcf", "colony", "bayescan", "gsisim", "hierfstat", "hzar", "ldna",
    "pcadapt", "related", "stockr", "structure", "arlequin", "arrow.parquet")
)
}
\description{
Generate a filename object
}
\keyword{internal}
