% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr.R
\name{generate_squeleton_folders}
\alias{generate_squeleton_folders}
\title{generate_squeleton_folders}
\usage{
generate_squeleton_folders(
  fp = 0L,
  path.folder = NULL,
  interactive.filter = TRUE,
  ...
)
}
\description{
Generate squeleton folders
}
\keyword{internal}
