% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_reconstruction.R
\name{haplotype_reconstruction}
\alias{haplotype_reconstruction}
\title{haplotype_reconstruction}
\usage{
haplotype_reconstruction(data, parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{A tidy genomic data frame or another genomic object supported by
the calling function.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\description{
Reconstruct haplotypes
}
\keyword{internal}
