% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Nucleotide diversity}
\usage{
pi(
  data,
  read.length = NULL,
  parallel.core = parallel::detectCores() - 1,
  path.folder = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{(4 options) A file or object generated by radr:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Use \code{\link[genometranslator]{read_genome}} to import supported formats
and \code{\link[genometranslator]{tidy_genome}} when a tidy table is needed.}

\item{read.length}{(integer, optional) The length in nucleotide of your reads.
By default it is estimated from the data using the column \code{COL}.
Default: \code{read.length = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{path.folder}{(path, optional) By default will print results in the working directory.
Default: \code{path.folder = NULL}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}
}
\value{
The function returns a list with the function call and:
\enumerate{
\item $pi.individuals: the pi estimated for each individual
\item $pi.populations: the pi statistics estimated per populations and overall.
\item $boxplot.pi: showing the boxplot of Pi for each populations and overall.
}
use $ to access each #' objects in the list.
}
\description{
Calculates the nucleotide diversity (Nei & Li, 1979).

To get an estimate with the consensus reads, use the
function \emph{summary_haplotypes} found in the package
\href{https://github.com/thierrygosselin/stackr}{stackr}. The estimate in
\emph{summary_haplotypes} integrates the consensus markers found in
\href{http://catchenlab.life.illinois.edu/stacks/}{STACKS}
populations.haplotypes.tsv file.
Both radr and stackr functions requires \code{stringdist} package.

The \code{read.length} argument below is used directly in the calculations.
To be correctly estimated, the reads obviously need to be of identical size...
}
\note{
Thanks to Anne-Laure Ferchaud for very useful comments on previous version
of this function.
}
\examples{
\dontrun{
require(stringdist)
# The simplest way to run the function:
pi.sum <- radr::pi(data = "brook.charr.gds")
}
}
\references{
Nei M, Li WH (1979)
Mathematical model for studying genetic variation in terms of
restriction endonucleases.
Proceedings of the National Academy of Sciences of
the United States of America, 76, 5269–5273.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
