% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_density_distribution_maf}
\alias{plot_density_distribution_maf}
\title{Figure density distribution of minor allele frequency (MAF)
summary statistics.}
\usage{
plot_density_distribution_maf(
  data,
  maf.group,
  aes.colour = ggplot2::aes(y = ..scaled.., color = POP_ID),
  adjust.bin = 1,
  x.title
)
}
\arguments{
\item{data}{sumstats or tidy vcf summarised files.}

\item{maf.group}{The GGPLOT2 aes (e.g. aes(x = FREQ_ALT, na.rm = F)).}

\item{aes.colour}{GGPLOT2 aesthetics colour,
e.g. aes(y = ..scaled.., color = POP_ID) or aes(y = ..count.., color = POP_ID)}

\item{adjust.bin}{Adjust GGPLOT2 bin size (0 to 1).}

\item{x.title}{Title of the x-axis. e.g. "MAF distribution"}
}
\description{
Create density distribution of MAF summary statistics.
}
\details{
turn off the legend using \code{fig + theme(legend.position = "none")} or
zoom in a section with 'fig + coord_cartesian(xlim = c(0, 0.1), ylim = c(0, 1))'.
Save the figure with : \code{ggsave("figure name.pdf", width = 40, height = 20, dpi = 600, units = "cm", useDingbats = F)}.
}
\seealso{
\link[genometranslator]{tidy_genome}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
\keyword{internal}
