% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_alleles.R
\name{private_alleles}
\alias{private_alleles}
\title{Find private alleles}
\usage{
private_alleles(data, strata = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A tidy genomic data frame or another genomic object supported by
the calling function.}

\item{strata}{(path or object) The strata file or object.
Additional documentation is available in \code{\link[genometranslator]{read_strata}}.
Use that function to whitelist/blacklist populations/individuals.
Option to set \code{pop.levels/pop.labels} is also available.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = FALSE}.}
}
\value{
A list with an object highlighting private alleles by markers and strata and
a second object with private alleles summarized by strata.
}
\description{
The function highlight private alleles by strata, using a GDS or tidy file or object.
}
\examples{
\dontrun{
corals.private.alleles.by.pop <- radr::private_alleles(data = tidy, strata = strata.pop)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
