% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr.R
\name{radr_function_header}
\alias{radr_function_header}
\title{radr_function_header}
\usage{
radr_function_header(f.name = NULL, start = TRUE, verbose = TRUE)
}
\description{
Generate function header
}
\keyword{internal}
