% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr.R
\name{radr_results_message}
\alias{radr_results_message}
\title{radr_results_message}
\usage{
radr_results_message(
  rad.message = NULL,
  filters.parameters,
  internal = FALSE,
  verbose = TRUE
)
}
\description{
Message printed at the end of most radr functions
}
\keyword{internal}
