% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_genomic_data.R
\name{summarise_genomic_data}
\alias{summarise_genomic_data}
\title{Summarise genomic data}
\usage{
summarise_genomic_data(data, path.folder = NULL, digits = 4, verbose = TRUE)
}
\arguments{
\item{data}{Tidy genomic data or a Genomic Data Structure (GDS) file or
object:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Use \code{\link[genometranslator]{read_genome}} to import supported formats
and \code{\link[genometranslator]{tidy_genome}} when a tidy table is needed.}

\item{path.folder}{(path, optional) By default will print results in the working directory.
Default: \code{path.folder = NULL}.}

\item{digits}{(integer, optional). Default: \code{digits = 4}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}
}
\description{
Calculate summary statistics from tidy genomic data or a GDS
file. Statistics are calculated by population and marker and include REF and
ALT allele frequencies, observed and expected heterozygosity, and the
inbreeding coefficient (FIS).
}
