% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_biallelic_problems.R
\name{detect_biallelic_problems}
\alias{detect_biallelic_problems}
\title{Detect biallelic problems}
\usage{
detect_biallelic_problems(
  data,
  verbose = TRUE,
  parallel.core = parallel::detectCores() - 1
)
}
\arguments{
\item{data}{A tidy data frame object in the global environment or
a tidy data frame in wide or long format in the working directory.
The tidy dataset needs a column with nucleotide information \code{GT_VCF_NUC},
usually this is automatically generated by \pkg{radr}.
\emph{How to get a tidy data frame ?}
Look into \pkg{genometranslator} \code{\link[genometranslator]{tidy_genome}}.}

\item{verbose}{(optional, logical) \code{verbose = TRUE} to be chatty
during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
Several info
}
\description{
Detect the number of alleles/nucleotides per markers.
Sometimes, a dataset might have 3 alleles at a SNP, is this biological or artifactual ?
This function helps to resolve this, by highlighting markers with this potential
problem, so that user can further look at the origin of the phenomenon.
The function can also split datasets in biallelic/multiallelic datasets.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
