% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_het_outliers.R
\name{detect_het_outliers}
\alias{detect_het_outliers}
\title{Detect heterozygotes outliers and estimate miscall rate}
\usage{
detect_het_outliers(
  data,
  nreps = 2000,
  burn.in = NULL,
  strata = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A GDS filename or open \code{SeqVarGDSClass} object.}

\item{nreps}{(integer, optional) The number of MCMC sweeps to do.
Default: \code{nreps = 2000}.}

\item{burn.in}{(integer, optional) The number of MCMC burn-in reps.
With default, during execution, you will be asked to enter the nuber of burn-in.
For this, a plot showing the heterozygote miscall rate for all
the MCMC sweeps will be printed. This plot will help pinpoint the
number of burn-in. The remaining MCMC sweeps will be used
to average the heterozygote miscall rate.
e.g. of common value \code{burn.in = 500}.
Default: \code{burn.in = NULL}.}

\item{strata}{Optional strata file or object containing individual and group
information. Default: \code{strata = NULL}.}

\item{filename}{Optional prefix used for generated output files.
Default: \code{filename = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
A folder generated automatically with date and time,
the file \code{het.summary.tsv} contains the summary statistics. The file
\code{markers.genotypes.boundaries.pdf} is the plot with boundaries.
The overall genotyping and heterozygotes miscall rate is writen in the file
\code{overall_error_rate.tsv}.
The function also returns a list inside the global environment with
8 objects:

\enumerate{
\item input the input data, cleaned if filters were used during import.
\item outlier.summary a list with a tibble and plot of genotypes frequencies
and boundaries (also written in the folder).
\item summary.alt.allele a tibble summarizing the number of markers with:
\itemize{
\item no homozygote for the alternate allele (NO_HOM_ALT)
\item no heterozygote genotype (NO_HET)
\item one homozygote for the alternate allele(ONE_HOM_ALT)
\item one heterozygote genotype (ONE_HET)
\item one homozygote for the alternate allele only (ONE_HOM_ALT_ONLY)
\item one heterozygote genotype only (ONE_HET_ONLY)
\item one homozygote for the alternate allele and one heterozygote genotype only (ONE_HOM_ALT_ONE_HET_ONLY)
}
\item m.nreps A tibble with the heterozygote miscall rate for each MCMC replicate
\item overall.genotyping.error.rate The overall genotyping error rate
\item overall.m The overall heterozygote miscall rate
\item simmed_genos The simulated genotypes
}

The statistics are summarized per population and overall,
the grouping is found in the last column called \code{POP_ID}.
}
\description{
Explore departure from H-W equilibrium in bi-allelic RADseq data.
Highlight excess of homozygotes present in numeros RADseq studies.
The function estimate the genotyping error rate and heterozygote miscall rate.
The model focus on heterozygotes being
incorrectly called as homozygotes. See details below for more info.
}
\details{
\strong{Before using the function:}
\enumerate{
\item Don't use raw RADseq data, this function will work best with filtered data
\item Remove duplicate \code{\link[radr]{detect_duplicate_genomes}}.
\item Remove mixed samples \code{\link[radr]{detect_mixed_genomes}}.
\item Look at other filters in radr package...
}

\strong{During import:}

By default the function will keep only polymorphic markers and markers common
between all populations. If you supply a tidy data frame or a \code{.rad} file,
the function skip all the filters, pop selection, etc. It will however scan and
remove monomorphic markers automatically.

\strong{Keep track of the data:}

Use the argument filename to write the imported (and maybe further filtered)
tidy genomic data set inside the folder. The filename will be automatically
appended \code{.rad} to it. This file can be used again directly inside this
function and other radr functions. See
\code{\link[genometranslator]{read_genome}}.
}
\examples{
\dontrun{
het.prob <- radr::detect_het_outliers(
data = "tuna.vcf", strata = "tuna.strata.tsv", nreps = 2000)
}
}
\author{
Eric Anderson \email{eric.anderson@noaa.gov} and
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
