% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_paralogs.R
\name{detect_paralogs}
\alias{detect_paralogs}
\title{Detect paralogs}
\usage{
detect_paralogs(
  data,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{(2 options) A Genomic Data Structure (GDS) file or object
generated by radr.

\emph{How to get GDS data ?}
Use \code{\link[genometranslator]{read_genome}} to import supported formats
and \code{\link[genometranslator]{tidy_genome}} when a tidy table is needed.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
The function returns the GDS data inside the global environment.
In the directory generated with \code{detect_paralogs} in it, the function
writes a tibble with the summary and the HD plot.

The tibble as 1 line per SNPs/MARKERS and columns are:
\itemize{
\item VARIANT_ID: the GDS SNP id, to have all the markers metadata use
\code{\link{extract_markers_metadata}}.
\item NUMBER_HET: the number of heterozygote genotype.
\item NUMBER_MISSING: the number of missing genotype.
\item NUMBER_GENOTYPED: the number of individual genotyped for the marker.
\item TOTAL: the total number of sample for this marker.
\item DEPTH_REF: the sum of the REF allele depth/coverage.
\item DEPTH_ALT: the sum of the ALT allele depth/coverage.
\item MISSING_PROP: Missing genotype proportion.
\item HET_PROP: the proportion if heterozygote sample.
\item TOTAL_COUNTS: Total coverage/counts for the marker.
\item RATIO: the allele ratio.
\item Z_SCORE: the z-score (deviation in read ratio).
}
}
\description{
The method uses the proportion of heterozygotes
and deviations in read ratios described in McKinney et al. 2017 to highlight
potential paralogous markers. Additionally, the function produces the HD
figure with marker's missingness accounted for.

\emph{Note:} Violating Assumptions/Prerequisites (see below) can lead to
false positive or the absence of detection of paralogous markers.
}
\section{Assumptions/Prerequisites}{

\enumerate{
\item \strong{read the paper:}
Read \href{https://onlinelibrary.wiley.com/doi/abs/10.1111/1755-0998.12613}{Garrett McKinney}
paper, twice. It's good.
Reference is below.
\item \strong{Depth/Coverage information:} The function requires read depth
information for each alleles (REF/ALT). This info will not be found in genepop
or structure file format. You usually need a VCF file for this.
\emph{stacks, ANGSD, GATK, platypus, ipyrad, DArT, VCFTools, PLINK and freebayes}
are software producing the required output.
\item \strong{Batch effect:} You don't have any batch effect, e.g. with
sequencer id, library, sampling sites, tissue preservation,
person behind the wet-lab bench conducting the DNA extraction
(operating the pipette, etc.).
\item \strong{Normalization:} Any form of normalization, before or after
sequencing, as the potential to bias paralogs discovery.
This include: combining lanes or chips to increase individual coverage,
normalizing individuals FASTQ files, and of course normalizing read depth.
\item \strong{Identity-by-Missingness:} Absence of artifactual pattern
of missingness (\href{https://github.com/thierrygosselin/grur}{see missing visualization})
\item \strong{Low genotyping error rate:} see \code{\link{detect_het_outliers}}
and \href{https://github.com/eriqande/whoa}{whoa}. You can also use:
\code{\link{filter_hwe}} with very low threshold to discard markers with
obvious genotyping errors.
\item \strong{Low heterozygosity miscall rate:} see \code{\link{detect_het_outliers}} and
\href{https://github.com/eriqande/whoa}{whoa}.
\item \strong{Absence of pattern of heterozygosity driven by missingness:}
see \code{\link{detect_mixed_genomes}}. Don't use this function if the
individual heterozygosity is not under control. If your data is not uniform
inside this function, don't expect much of \code{detect_paralogs}.
\item Use unfiltered dataset for fun only.
\item Use this function preferably on filtered data, check
\code{\link[radr]{explore_genomes}}.
\item \strong{Sex-biased sampling and sex markers} will generate
false-positive with this function.
Use \code{\link{sexy_markers}} if you have sexe information.
}
}

\examples{
\dontrun{
# require(httr)
# using the Chinook VCF dataset in dryad, from McKinney et al. 2017 paper:
writeBin(object = httr::content(x = httr::GET(
url = "https://datadryad.org/bitstream/handle/10255/dryad.123464/
Chinook_sequenceReads.vcf?sequence=1"),
as = "raw"), con = "chinook.vcf")

# Import VCF and run the function
chinook.paralogs <- genometranslator::read_vcf(data = "chinook.vcf", vcf.stats = FALSE) \%>\%
    radr::detect_paralogs(data = .)
}
}
\references{
McKinney GJ, Waples RK, Seeb LW, Seeb JE (2017)
Paralogs are revealed by proportion of heterozygotes and deviations in read
ratios in genotyping-by-sequencing data from natural populations.
Molecular Ecology Resources, 17, 656-669.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
