% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_ma.R
\name{diagnostic_ma}
\alias{diagnostic_ma}
\title{MA diagnostic}
\usage{
diagnostic_ma(data, group.rank, filename = NULL)
}
\arguments{
\item{data}{A file in the working directory or object in the global environment
in wide or long (tidy) formats. To import, the function uses
\href{https://github.com/thierrygosselin/radr}{radr}
\code{\link[radr]{read_genome}}.
\emph{See details of this function for more info}.}

\item{group.rank}{(Number) The number of group to class the MAF.}

\item{filename}{(optional) Name of the file written to the working directory.}
}
\description{
Minor Allele diagnostic, help choose a filter threshold.
}
\details{
Highly recommended to look at the distribution of MAF
}
\examples{
\dontrun{
problem <- radr::diagnostic_ma(
data = tidy.salmon.data, group.rank = 10)
}
}
\seealso{
\link{filter_ma}
}
\keyword{internal}
