% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_blacklist_genotypes.R
\name{filter_blacklist_genotypes}
\alias{filter_blacklist_genotypes}
\title{Filter dataset with blacklist of genotypes}
\usage{
filter_blacklist_genotypes(data, blacklist.genotypes, verbose = TRUE, ...)
}
\arguments{
\item{data}{(4 options) A file or object generated by radr:
\itemize{
\item tidy data
\item Genomic Data Structure (GDS)
}

\emph{How to get GDS and tidy data ?}
Use \code{\link[genometranslator]{read_genome}} to import supported formats
and \code{\link[genometranslator]{tidy_genome}} when a tidy table is needed.}

\item{blacklist.genotypes}{A required blacklist object or filepath containing
genotype records to remove or mask.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\description{
Filter dataset with blacklist of genotypes.

This function allows to blacklist/erase/mask genotypes.

Used internally in \href{https://github.com/thierrygosselin/radr}{radr}
and might be of interest for users.
}
\section{Life cycle}{


This function arguments will be subject to changes. Currently the function uses
erase.genotypes, but using the \code{dots-dots-dots ...} arguments allows to
pass \code{erase.genotypes and masked.genotypes}. These arguments do exactly
the same thing and only one can be used.
}

\examples{
\dontrun{
data <- radr::filter_blacklist_genotypes(
    data = data, blacklist.geno = "blacklist.geno.tsv"
    )
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
