% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dart_reproducibility.R
\name{filter_dart_reproducibility}
\alias{filter_dart_reproducibility}
\title{Filter data based on DArT reproducibility statistics}
\usage{
filter_dart_reproducibility(
  data,
  interactive.filter = TRUE,
  filter.reproducibility = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A tidy genomic data frame or another genomic object supported by
the calling function.}

\item{interactive.filter}{Logical indicating whether an interactive filtering
session may display diagnostics and ask for thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{filter.reproducibility}{(double, character) This is best decided after viewing the figures.
Usually values higher than 0.95 are not uncommon.
The value can also be character: \code{filter.reproducibility = "outliers"}.
Using this, will remove outlier markers using the lower outlier statistics.
Default: \code{filter.reproducibility = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical indicating whether progress messages are emitted.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $whitelist.markers
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes markers below a certain threshold.
Based on the repoducibility column found in DArT files.

\strong{Filter targets}: Markers

\strong{Statistics}: Reproducibility (established by DArT)
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the reproducibility value:

Step 1. Visualization using a box plot

Step 2. Choose the filtering threshold
}
\examples{
\dontrun{
spotted.cod <- genometranslator::read_dart(
    data = "Combined_1514and1614_SNP_80Callrate.csv",
    strata = "strata.dart.spotted.cod.tsv"
)
turtle.filtered <- radr::filter_dart_reproducibility(
    data = spotted.cod,
    filter.reproducibility = 0.97
)
}
}
