% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_het.R
\name{filter_het}
\alias{filter_het}
\title{Heterozygosity filter}
\usage{
filter_het(
  interactive.filter = TRUE,
  data,
  strata = NULL,
  ind.heterozygosity.threshold = NULL,
  het.approach = c("SNP", "overall"),
  het.threshold = 1,
  het.dif.threshold = 1,
  outlier.pop.threshold = 1,
  helper.tables = FALSE,
  coverage.info = FALSE,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{interactive.filter}{(optional, logical) Do you want the filtering session to
be interactive. With the default, the user is asked to see figures of
distribution before making decisions for filtering with heterozygosity statistics.
Default: \code{interactive.filter = TRUE}.}

\item{data}{A GDS filename or open \code{SeqVarGDSClass} object.}

\item{strata}{Optional strata file or object containing individual and group
information. Default: \code{strata = NULL}.}

\item{ind.heterozygosity.threshold}{(string, double, optional)
Blacklist individuals based on observed heterozygosity (averaged across markers).

The string contains 2 thresholds values (min and max).
The values are proportions (0 to 1), where 0 turns off the min threshold and
1 turns off the max threshold.
Individuals with mean observed heterozygosity higher (>) or lower (<)
than the thresholds will be blacklisted.

A value of \code{NULL} turns off filtering while retaining the plots and
heterozygosity tables.
Default: \code{ind.heterozygosity.threshold = NULL}.}

\item{het.approach}{(Character string). First value, \code{"SNP"} or
\code{"haplotype"}. The haplotype approach considers the statistic
consistency on the read (locus/haplotype). The major difference:
the haplotype approach results in blacklisting the entire locus/haplotype
with all the SNPs on the read.
With the SNP approach, SNPs are independently analyzed and blacklisted.
The second value, will use the statistics by population \code{"pop"} or
will consider the data overall, as 1 large group \code{"overall"}.
Default: \code{het.approach = c("SNP", "overall")}.}

\item{het.threshold}{Number Biallelic markers usually max 0.5. But departure
from this value is common. The higher the proportion threshold, the more relaxed
the filter is. With default there is no filtering.
Default: \code{het.threshold = 1}.}

\item{het.dif.threshold}{Number (0 - 1). For \code{het.approach = "haplotype"} only.
You can set a threshold for the difference in het along your read.
Set the number your willing to tolerate on the same read/haplotype.
e.g. if you have 2 SNP on a read/haplotype and on as a het of 0.9 and the other
0.1 and you set \code{het.dif.threshold = 0.3}, this markers will be blacklisted.
You should strive to have similar statistics along short read like RAD.
The higher the proportion threshold, the more relaxed the filter is.
With default, there is no filtering and all the range of differences are allowed.
Default: \code{het.dif.threshold = 1}.}

\item{outlier.pop.threshold}{(integer, optional) Useful to incorporate problematic
populations dragging down polymorphism discovery, but still wanted for analysis.
Use this threshold to allow variance in the number of populations passing
the thresholds described above.
e.g. with \code{outlier.pop.threshold = 2},
you tolerate a maximum of 2 populations failing the
\code{het.threshold} and/or \code{het.dif.threshold}.
Manage outlier markers, individuals and populations
downstream with blacklists and whitelists produced by the function.
See details for more information.
Default: \code{outlier.pop.threshold = 1}.}

\item{helper.tables}{(logical) Output tables that show
the number of markers blacklisted or whitelisted based on a series of
automatic thresholds to guide decisions. When \code{interactive.filter == TRUE},
helper tables are written to the directory.
Default: \code{helper.tables = FALSE}.}

\item{coverage.info}{(optional, logical) Use
\code{coverage.info = TRUE}, if you want to visualize
the relationshio between locus coverage and locus observed heterozygosity
statistics. Coverage information is required (e.g. in a vcf file...).
Default: \code{coverage.info = FALSE}.}

\item{filename}{(optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in
the folder created in the working directory. The file extension appended to
the \code{filename} provided is \code{.rad}.
Default: \code{filename = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
The function returns inside the global environment a list with
15 objects, the objects names are found by using \code{names(your.list.name)}:

\enumerate{
\item filtered tidy data frame: \code{$tidy.filtered.het}
\item whitelist of markers:\code{$whitelist.markers}
\item the strata:\code{$strata}
\item the filters parameters used:\code{$filters.parameters}
\item the individual's heterozigosity:\code{$individual.heterozigosity}
\item the heterozygosity statistics per populations and overall:\code{$heterozygosity.statistics}
\item the blacklisted individuals based on the individual's heterozigosity:\code{$blacklist.ind.het}
\item a list containing the helper tables:\code{$helper.table.het}
\item the boxplot of individual observed heterozygosity:\code{$individual.heterozygosity.boxplot}
\item the manhattan plot of individual heterozygosity:\code{$individual.heterozygosity.manhattan.plot}
\item the boxplot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.boxplot}
\item the density plot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.density.plot}
\item the manhattan plot of observed heterozygosity averaged across markers and pop:\code{$markers.pop.heterozygosity.manhattan.plot}
\item the relationship between the number of SNPs on a locus and locus observed heterozygosity statistics:\code{$snp.per.locus.het.plot}
\item the relationship between locus coverage (read depth) and locus observed heterozygosity statistics:\code{$het.cov.plot}
\item whitelist of markers:\code{$whitelist.markers}
\item whitelist of markers:\code{$whitelist.markers}
}

In the working directory, output is conditional to \code{interactive.filter} argument
}
\description{
Observed Heterozygosity based filtering.
The filter arguments of \code{filter_het} allows you to test rapidly
if departure from realistic expectations of heterozygosity statistics
are a problem in downstream analysis.

\strong{Filter target:} Individuals and/or markers, depending on the selected
arguments.
\enumerate{
\item Highlight outliers individual's observed heterozygosity for a quick
diagnostic of mixed samples or poor polymorphism discovery. The statistic is
also contrasted with missing data to help differentiate problems.
(also computed in \link{detect_mixed_genomes})

\item The observed heterozygosity in the dataset: an assembly artefact,
a genotyping problem, a problem of population groupings
or a reliable signal of biological polymorphism?
Detect assembly artifact or genotyping problem (e.g. under/over-splitting loci)
by looking at marker's observed heterozygosity statistics by population
or overall.

\item Use haplotype or snp level statistics. When the haplotype approach
is selected, consistencies of heterozygosity statistics along the read are
highlighted.

\item Interactive approach help by visualizing the data before making
a decision on thresholds.
}
}
\details{
\strong{Interactive version}

There are 4 steps in the interactive version to visualize and filter
the data based heterozygosity statistics:

Step 1. Individual's observed heterozygosity: outliers that might represent mixed samples
Step 2. Blacklist outliers based on a proportion threshold of mean observed heterozygosity
Step 3. Observed heterozygosity statistics per populations and overall
Step 4: Blacklist markers based on observed heterozygosity

\strong{outlier.pop.threshold}

If your a regular radr user, you've seen the \code{pop.num.threshold}.
\code{outlier.pop.threshold}, is different and requires more thinking,
because the number of populations genotyped potentially vary across markers,
which makes the use of \code{pop.num.threshold} less optimal.
e.g. If only 4 populations out of 10 are genotyped, for a marker you want
to keep, using \code{pop.num.threshold = 0.6} will lead to unwanted results
and inconsistensis.
It's easier to tolerate outliers with this new approach:
\code{outlier.pop.threshold = 2}.

\strong{Individual observed heterozygosity (averaged across markers):}
To help discard an individual based on his observed heterozygosity
(averaged across markers),
use the manhanttan plot to:
\enumerate{
\item contrast the individual with population and overall samples.
\item visualize the impact of missigness information (based on population or
overall number of markers) and the individual observed heterozygosity. The
larger the point, the more missing genotypes.
}
\strong{Outlier above average:}
\itemize{
\item potentially represent two samples mixed together (action: blacklist), or...
\item a sample with more sequecing effort (point size small): did you merge your replicates fq files ? (action : keep and monitor)
\item a sample with poor sequencing effort (point size large) where the genotyped markers are
all heterozygotes, verify this with missingness (action: discard)
}
In all cases, if there is no bias in the population sequencing effort,
the size of the point will usually be "average" based on the population or
overall number of markers.

You can visualize individual observed heterozygosity, choose thresholds and
then visualize, choose thresholds and filter markers based on observed
heterozygosity in one run with: \pkg{radr} \code{\link{filter_het}}.

\strong{Outlier below average:}
\itemize{
\item A point with a size larger than the population or overall average (= lots of missing):
the poor polymorphism discovery of the sample is probably the result of bad
DNA quality, a bias in sequencing effort, etc. (action: blacklist)
\item A point with a size that looks average (not much missing):
this sample requires more attention (action: blacklist) and conduct more tests.
e.g. for biallelic data, look for coverage imbalance between ALT/REF allele.
At this point you need to distinguish between an artifact of poor polymorphism discovery
or a biological reason (highly inbred individual, etc.).
}
}
\seealso{
\link{plot_density_distribution_het}
}
