% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radr.R
\name{filter_parameters}
\alias{filter_parameters}
\title{Track radr filtering parameters}
\usage{
filter_parameters(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[genometranslator:genome_parameters]{genometranslator::genome_parameters()}}.}
}
\description{
A package-facing wrapper around
\code{\link[genometranslator:genome_parameters]{genometranslator::genome_parameters()}} that maintains the common genomic
operation history during radr filtering.
}
\keyword{internal}
