% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snp_number.R
\name{filter_snp_number}
\alias{filter_snp_number}
\title{Filter SNP number per locus/read}
\usage{
filter_snp_number(
  data,
  strata = NULL,
  interactive.filter = TRUE,
  filter.snp.number = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A tidy genomic data frame or another genomic object supported by
the calling function.}

\item{strata}{Optional strata file or object containing individual and group
information. Default: \code{strata = NULL}.}

\item{interactive.filter}{Logical indicating whether an interactive filtering
session may display diagnostics and ask for thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{filter.snp.number}{(integer) This is best decided after viewing the figures.
If the argument is set to 2, locus with 3 and more SNPs will be blacklisted.
Default: \code{filter.snp.number = NULL}.}

\item{filename}{(optional) Name of the filtered tidy data frame file
written to the working directory (ending with \code{.tsv})
Default: \code{filename = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $snp.number.markers
\item $number.snp.reads.plot
\item $whitelist.markers
\item $tidy.filtered.snp.number
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes outlier markers
with too many SNP number per locus/read.
The data requires snp and locus information (e.g. from a VCF file).
Having a higher than "normal" SNP number is usually the results of
assembly artifacts or bad assembly parameters.
This filter is population-agnostic, but still requires a strata
file if a vcf file is used as input.

\strong{Filter targets}: Markers

\strong{Statistics}: The number of SNPs per locus.
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the number of SNP on the read/locus:

Step 1. SNP number per read/locus visualization

Step 2. Choose the filtering thresholds
}
\examples{
\dontrun{
turtle.outlier.snp.number <- radr::filter_snp_number(
data = "turtle.vcf",
strata = "turtle.strata.tsv",
max.snp.number = 4,
filename = "tidy.data.turtle.tsv"
)

tidy.data <- turtle.outlier.snp.number$tidy.filtered.snp.number

#Inside the same list, to isolate the markers blacklisted:
blacklist <- turtle.outlier.snp.number$blacklist.markers

}
}
