% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snp_position_read.R
\name{filter_snp_position_read}
\alias{filter_snp_position_read}
\title{Filter markers/SNP based on their position on the read}
\usage{
filter_snp_position_read(
  data,
  strata = NULL,
  interactive.filter = TRUE,
  filter.snp.position.read = NULL,
  filename = NULL,
  parallel.core = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A tidy genomic data frame or another genomic object supported by
the calling function.}

\item{strata}{Optional strata file or object containing individual and group
information. Default: \code{strata = NULL}.}

\item{interactive.filter}{Logical indicating whether an interactive filtering
session may display diagnostics and ask for thresholds.
Default: \code{interactive.filter = TRUE}.}

\item{filter.snp.position.read}{(character)
Options are: \code{"outliers", "q75", "iqr", c(min value,max value)}.
For a safe and conservative
value, use \code{"outliers"}, this will remove SNPs with outlier position on
the reads.
Default: \code{filter.snp.position.read = NULL}.}

\item{filename}{(optional) Name of the filtered tidy data frame file
written to the working directory (ending with \code{.tsv})
Default: \code{filename = NULL}.}

\item{parallel.core}{Number of workers available for parallel operations.
Default: \code{parallel.core = parallel::detectCores() - 1}.}

\item{verbose}{Logical. Display progress messages.
Default: \code{verbose = TRUE}.}

\item{...}{Additional arguments passed to lower-level screening or filtering
functions.}
}
\value{
A list in the global environment with 6 objects:
\enumerate{
\item $snp.number.markers
\item $number.snp.reads.plot
\item $whitelist.markers
\item $tidy.filtered.snp.number
\item $blacklist.markers
\item $filters.parameters
}

The object can be isolated in separate object outside the list by
following the example below.
}
\description{
This filter removes markers/SNPs based on their position on the read.
The data requires snp, locus and col information (e.g. from a VCF file).

The impact of assembly artifacts can be tested in downstream analysis with
the whitelist and blacklist generated by this function.

\strong{Filter target:} Markers.

\strong{Statistics}: The position of the SNPs on the read.
}
\details{
\strong{Interactive version}

There are 2 steps in the interactive version to visualize and filter
the data based on the number of SNP on the read/locus:

Step 1. SNP number per read/locus visualization

Step 2. Choose the filtering thresholds
}
\examples{
\dontrun{
turtle <- radr::filter_snp_position_read(
data = "turtle.vcf",
strata = "turtle.strata.tsv",
filter.snp.position.read = "outliers",
filename = "tidy.data.turtle.tsv"
)
}
}
